Make `QuillProm` be a configured value too.
This is done because we can't count on configuration having been performed
by the time the `quill-util` module is loaded.
