+++ local-modules/@bayou/config-client/Editor.js
@@ -19,6 +19,15 @@ export default class Editor extends UtilityClass {
   }
 
   /**
+   * {QuillProm} A subclass of `Quill` which supports promise-chain event
+   * access, as exemplified in the default implementation of this configuration
+   * point in `config-client-default`.
+   */
+  static get QuillProm() {
+    return use.Editor.QuillProm;
+  }
+
+  /**
    * Performs any webpage-global setup needed in order for the page to host one
    * or more editors. This is called exactly once per page load, early during
    * initialization and specifically _before_ any programmatic DOM manipulation
