+++ local-modules/@bayou/quill-util/index.js
@@ -2,14 +2,14 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
+import PromSubclasser from './PromSubclasser';
 import QuillEvents from './QuillEvents';
 import QuillGeometry from './QuillGeometry';
-import QuillProm from './QuillProm';
 import QuillUtil from './QuillUtil';
 
 export {
+  PromSubclasser,
   QuillEvents,
   QuillGeometry,
-  QuillProm,
-  QuillUtil,
+  QuillUtil
 };
